refactor: Split wasm-bindings/lib.rs into modular structure
- Created 6 logical modules by functionality:
  * types.rs - Result types and schemas (50 lines)
  * database.rs - Database struct and constructor (28 lines)
  * execute.rs - DDL/DML execution (71 lines)
  * query.rs - SELECT query execution and JSON formatting (80 lines)
  * schema.rs - Schema introspection (list_tables, describe_table) (42 lines)
  * examples.rs - Example database loaders (58 lines)
  * tests.rs - All unit tests (368 lines)
- Converted lib.rs to module imports and re-exports (12 lines)
- All tests passing
- No functional changes, pure refactoring

Reduces wasm-bindings/lib.rs from 662 lines to ~12 lines + 7 focused modules
