+++ crates/wasm-bindings/src/lib.rs
@@ -1,662 +1,12 @@
-use serde::{Deserialize, Serialize};
 use wasm_bindgen::prelude::*;
+
+mod modules;
+
+pub use modules::{ColumnInfo, Database, ExecuteResult, QueryResult, TableSchema};
 
 /// Initializes the WASM module and sets up panic hooks
 #[wasm_bindgen(start)]
 pub fn init_wasm() {
     // Better error messages in browser console
     console_error_panic_hook::set_once();
-}
-
-/// Result of a query execution
-#[derive(Serialize, Deserialize)]
-#[wasm_bindgen(getter_with_clone)]
-pub struct QueryResult {
-    /// Column names
-    pub columns: Vec<String>,
-    /// Row data as JSON strings
-    pub rows: Vec<String>,
-    /// Number of rows
-    pub row_count: usize,
-}
-
-/// Result of an execute (DDL/DML) operation
-#[derive(Serialize, Deserialize)]
-#[wasm_bindgen(getter_with_clone)]
-pub struct ExecuteResult {
-    /// Number of rows affected (for INSERT, UPDATE, DELETE)
-    pub rows_affected: usize,
-    /// Success message
-    pub message: String,
-}
-
-/// Table column metadata
-#[derive(Clone, Serialize, Deserialize)]
-#[wasm_bindgen(getter_with_clone)]
-pub struct ColumnInfo {
-    /// Column name
-    pub name: String,
-    /// Data type (as string)
-    pub data_type: String,
-    /// Whether column can be NULL
-    pub nullable: bool,
-}
-
-/// Table schema information
-#[derive(Clone, Serialize, Deserialize)]
-#[wasm_bindgen(getter_with_clone)]
-pub struct TableSchema {
-    /// Table name
-    pub name: String,
-    /// Column definitions
-    pub columns: Vec<ColumnInfo>,
-}
-
-/// In-memory SQL database with WASM bindings
-#[wasm_bindgen]
-pub struct Database {
-    db: storage::Database,
-}
-
-#[wasm_bindgen]
-impl Database {
-    /// Creates a new empty database instance
-    #[wasm_bindgen(constructor)]
-    pub fn new() -> Database {
-        Database { db: storage::Database::new() }
-    }
-
-    /// Executes a DDL or DML statement (CREATE TABLE, INSERT, UPDATE, DELETE)
-    /// Returns a JSON string with the result
-    pub fn execute(&mut self, sql: &str) -> Result<JsValue, JsValue> {
-        // Parse the SQL
-        let stmt = parser::Parser::parse_sql(sql)
-            .map_err(|e| JsValue::from_str(&format!("Parse error: {:?}", e)))?;
-
-        // Execute based on statement type
-        match stmt {
-            ast::Statement::CreateTable(create_stmt) => {
-                executor::CreateTableExecutor::execute(&create_stmt, &mut self.db)
-                    .map_err(|e| JsValue::from_str(&format!("Execution error: {:?}", e)))?;
-
-                let result = ExecuteResult {
-                    rows_affected: 0,
-                    message: format!("Table '{}' created successfully", create_stmt.table_name),
-                };
-
-                serde_wasm_bindgen::to_value(&result)
-                    .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-            }
-            ast::Statement::DropTable(drop_stmt) => {
-                let message = executor::DropTableExecutor::execute(&drop_stmt, &mut self.db)
-                    .map_err(|e| JsValue::from_str(&format!("Execution error: {:?}", e)))?;
-
-                let result = ExecuteResult {
-                    rows_affected: 0,
-                    message,
-                };
-
-                serde_wasm_bindgen::to_value(&result)
-                    .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-            }
-            ast::Statement::Insert(insert_stmt) => {
-                executor::InsertExecutor::execute(&mut self.db, &insert_stmt)
-                    .map_err(|e| JsValue::from_str(&format!("Execution error: {:?}", e)))?;
-
-                let row_count = insert_stmt.values.len();
-                let result = ExecuteResult {
-                    rows_affected: row_count,
-                    message: format!(
-                        "{} row{} inserted into '{}'",
-                        row_count,
-                        if row_count == 1 { "" } else { "s" },
-                        insert_stmt.table_name
-                    ),
-                };
-
-                serde_wasm_bindgen::to_value(&result)
-                    .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-            }
-            ast::Statement::Select(_) => {
-                Err(JsValue::from_str("Use query() method for SELECT statements"))
-            }
-            _ => Err(JsValue::from_str(&format!(
-                "Statement type not yet supported in WASM: {:?}",
-                stmt
-            ))),
-        }
-    }
-
-    /// Executes a SELECT query and returns results as JSON
-    pub fn query(&self, sql: &str) -> Result<JsValue, JsValue> {
-        // Parse the SQL
-        let stmt = parser::Parser::parse_sql(sql)
-            .map_err(|e| JsValue::from_str(&format!("Parse error: {:?}", e)))?;
-
-        // Ensure it's a SELECT statement
-        let select_stmt = match stmt {
-            ast::Statement::Select(s) => s,
-            _ => return Err(JsValue::from_str("query() method requires a SELECT statement")),
-        };
-
-        // Execute the query with column metadata
-        let select_executor = executor::SelectExecutor::new(&self.db);
-        let result = select_executor
-            .execute_with_columns(&select_stmt)
-            .map_err(|e| JsValue::from_str(&format!("Execution error: {:?}", e)))?;
-
-        let columns = result.columns;
-        let rows = result.rows;
-
-        // Convert rows to JSON strings
-        let row_strings: Vec<String> = rows
-            .iter()
-            .map(|row| {
-                // Convert each SqlValue to a JSON-compatible representation
-                let json_values: Vec<serde_json::Value> = row
-                    .values
-                    .iter()
-                    .map(|v| match v {
-                        types::SqlValue::Integer(i) => serde_json::Value::Number((*i).into()),
-                        types::SqlValue::Smallint(i) => serde_json::Value::Number((*i).into()),
-                        types::SqlValue::Bigint(i) => serde_json::Value::Number((*i).into()),
-                        types::SqlValue::Float(f) => {
-                            serde_json::Number::from_f64(*f as f64)
-                                .map(serde_json::Value::Number)
-                                .unwrap_or(serde_json::Value::Null)
-                        }
-                        types::SqlValue::Real(f) => {
-                            serde_json::Number::from_f64(*f as f64)
-                                .map(serde_json::Value::Number)
-                                .unwrap_or(serde_json::Value::Null)
-                        }
-                        types::SqlValue::Double(f) => {
-                            serde_json::Number::from_f64(*f)
-                                .map(serde_json::Value::Number)
-                                .unwrap_or(serde_json::Value::Null)
-                        }
-                        types::SqlValue::Varchar(s) | types::SqlValue::Character(s) => {
-                            serde_json::Value::String(s.clone())
-                        }
-                        types::SqlValue::Boolean(b) => serde_json::Value::Bool(*b),
-                        types::SqlValue::Numeric(s)
-                        | types::SqlValue::Date(s)
-                        | types::SqlValue::Time(s)
-                        | types::SqlValue::Timestamp(s)
-                        | types::SqlValue::Interval(s) => serde_json::Value::String(s.clone()),
-                        types::SqlValue::Null => serde_json::Value::Null,
-                    })
-                    .collect();
-
-                serde_json::to_string(&json_values).unwrap_or_else(|_| "[]".to_string())
-            })
-            .collect();
-
-        let result = QueryResult { columns, rows: row_strings, row_count: rows.len() };
-
-        serde_wasm_bindgen::to_value(&result)
-            .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-    }
-
-    /// Lists all table names in the database
-    pub fn list_tables(&self) -> Result<JsValue, JsValue> {
-        let table_names: Vec<String> = self.db.list_tables();
-
-        serde_wasm_bindgen::to_value(&table_names)
-            .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-    }
-
-    /// Gets the schema for a specific table
-    pub fn describe_table(&self, table_name: &str) -> Result<JsValue, JsValue> {
-        let table = self
-            .db
-            .get_table(table_name)
-            .ok_or_else(|| JsValue::from_str(&format!("Table '{}' not found", table_name)))?;
-
-        let schema = &table.schema;
-        let columns: Vec<ColumnInfo> = schema
-            .columns
-            .iter()
-            .map(|col| ColumnInfo {
-                name: col.name.clone(),
-                data_type: format!("{:?}", col.data_type),
-                nullable: col.nullable,
-            })
-            .collect();
-
-        let table_schema = TableSchema { name: table_name.to_string(), columns };
-
-        serde_wasm_bindgen::to_value(&table_schema)
-            .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-    }
-
-    /// Loads the Employees example database (hierarchical org structure)
-    /// Demonstrates recursive queries with WITH RECURSIVE
-    pub fn load_employees(&mut self) -> Result<JsValue, JsValue> {
-        let sql = include_str!("../../../web-demo/examples/employees.sql");
-
-        // Execute the SQL file (contains CREATE TABLE and INSERT statements)
-        // Split by semicolons and execute each statement
-        for statement_sql in sql.split(';') {
-            let trimmed = statement_sql.trim();
-            if trimmed.is_empty() || trimmed.starts_with("--") {
-                continue; // Skip empty lines and comments
-            }
-
-            self.execute(trimmed)?;
-        }
-
-        let result = ExecuteResult {
-            rows_affected: 35, // 35 employees inserted
-            message: "Employees database loaded successfully (35 employees)".to_string(),
-        };
-
-        serde_wasm_bindgen::to_value(&result)
-            .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-    }
-
-    /// Loads the Northwind example database (classic sales/orders database)
-    /// Demonstrates JOINs, aggregates, and relational database concepts
-    pub fn load_northwind(&mut self) -> Result<JsValue, JsValue> {
-        let sql = include_str!("../../../web-demo/examples/northwind.sql");
-
-        // Execute the SQL file (contains CREATE TABLE and INSERT statements)
-        // Split by semicolons and execute each statement
-        for statement_sql in sql.split(';') {
-            let trimmed = statement_sql.trim();
-            if trimmed.is_empty() || trimmed.starts_with("--") {
-                continue; // Skip empty lines and comments
-            }
-
-            self.execute(trimmed)?;
-        }
-
-        let result = ExecuteResult {
-            rows_affected: 143, // Total rows across all tables
-            message: "Northwind database loaded successfully (5 tables, 143 rows)".to_string(),
-        };
-
-        serde_wasm_bindgen::to_value(&result)
-            .map_err(|e| JsValue::from_str(&format!("Serialization error: {:?}", e)))
-    }
-
-    /// Returns the version string
-    pub fn version(&self) -> String {
-        "nistmemsql-wasm 0.1.0".to_string()
-    }
-}
-
-impl Default for Database {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_database_creation() {
-        let db = Database::new();
-        assert_eq!(db.version(), "nistmemsql-wasm 0.1.0");
-    }
-
-    // Note: Tests that call JsValue-returning methods must use wasm-pack test --node
-    // The methods above work correctly but require a WASM runtime to test
-    // For now, we verify that the code compiles correctly
-
-    // ========== Examples Validation Tests ==========
-    // Auto-validation tests for web-demo examples
-    // These tests read examples.ts at compile time and validate each SQL query
-
-    /// Helper to execute SQL (handles multi-statement SQL like loading database schemas)
-    fn execute_sql(db: &mut storage::Database, sql: &str) -> Result<(), String> {
-        // Split by semicolons to handle multiple statements
-        for statement_text in sql.split(';') {
-            // Remove comment lines and trim
-            let cleaned: String = statement_text
-                .lines()
-                .filter(|line| !line.trim().is_empty() && !line.trim().starts_with("--"))
-                .collect::<Vec<&str>>()
-                .join("\n");
-
-            let trimmed = cleaned.trim();
-            if trimmed.is_empty() {
-                continue;
-            }
-
-            // Parse the SQL
-            let stmt = parser::Parser::parse_sql(trimmed)
-                .map_err(|e| {
-                    // Show first 100 chars of SQL for debugging
-                    let sql_preview = if trimmed.len() > 100 {
-                        format!("{}...", &trimmed[..100])
-                    } else {
-                        trimmed.to_string()
-                    };
-                    format!("Parse error in '{}': {:?}", sql_preview, e)
-                })?;
-
-            // Execute based on statement type
-            match stmt {
-                ast::Statement::CreateTable(create_stmt) => {
-                    executor::CreateTableExecutor::execute(&create_stmt, db)
-                        .map_err(|e| format!("CreateTable error: {:?}", e))?;
-                }
-                ast::Statement::DropTable(drop_stmt) => {
-                    executor::DropTableExecutor::execute(&drop_stmt, db)
-                        .map_err(|e| format!("DropTable error: {:?}", e))?;
-                }
-                ast::Statement::Insert(insert_stmt) => {
-                    executor::InsertExecutor::execute(db, &insert_stmt)
-                        .map_err(|e| format!("Insert error: {:?}", e))?;
-                }
-                ast::Statement::Select(select_stmt) => {
-                    let select_executor = executor::SelectExecutor::new(db);
-                    select_executor.execute(&select_stmt)
-                        .map_err(|e| format!("Select error: {:?}", e))?;
-                }
-                ast::Statement::Update(update_stmt) => {
-                    executor::UpdateExecutor::execute(&update_stmt, db)
-                        .map_err(|e| format!("Update error: {:?}", e))?;
-                }
-                ast::Statement::Delete(delete_stmt) => {
-                    executor::DeleteExecutor::execute(&delete_stmt, db)
-                        .map_err(|e| format!("Delete error: {:?}", e))?;
-                }
-                _ => return Err(format!("Unsupported statement type: {:?}", stmt)),
-            }
-        }
-        Ok(())
-    }
-
-    /// Helper to create the appropriate database based on name
-    /// Returns (Database, Option<load_error>)
-    fn setup_example_database(db_name: &str) -> (storage::Database, Option<String>) {
-        match db_name {
-            "northwind" => {
-                let mut db = storage::Database::new();
-                let northwind_sql = include_str!("../../../web-demo/examples/northwind.sql");
-                match execute_sql(&mut db, northwind_sql) {
-                    Ok(_) => (db, None),
-                    Err(e) => (storage::Database::new(), Some(format!("northwind DB load failed: {}", e))),
-                }
-            }
-            "employees" => {
-                let mut db = storage::Database::new();
-                let employees_sql = include_str!("../../../web-demo/examples/employees.sql");
-                match execute_sql(&mut db, employees_sql) {
-                    Ok(_) => (db, None),
-                    Err(e) => (storage::Database::new(), Some(format!("employees DB load failed: {}", e))),
-                }
-            }
-            // All other databases (empty, company, university, etc.) start as empty
-            // Examples using these databases create their own tables
-            _ => (storage::Database::new(), None),
-        }
-    }
-
-    /// Parse examples.ts to extract SQL queries for testing
-    /// Returns: Vec<(id, database, sql)>
-    fn parse_examples() -> Vec<(String, String, String)> {
-        let examples_ts = include_str!("../../../web-demo/src/data/examples.ts");
-
-        let mut examples = Vec::new();
-        let lines: Vec<&str> = examples_ts.lines().collect();
-
-        let mut i = 0;
-        while i < lines.len() {
-            // Look for example id
-            if let Some(id_line) = lines[i].strip_prefix("        id: '") {
-                if let Some(id_end) = id_line.find("'") {
-                    let id = &id_line[..id_end];
-
-                    // Look for database (within next 5 lines)
-                    let mut database = String::new();
-                    for j in (i + 1)..(i + 6).min(lines.len()) {
-                        if let Some(db_line) = lines[j].strip_prefix("        database: '") {
-                            if let Some(db_end) = db_line.find("'") {
-                                database = db_line[..db_end].to_string();
-                                break;
-                            }
-                        }
-                    }
-
-                    // Look for SQL (within next 15 lines)
-                    for j in (i + 1)..(i + 15).min(lines.len()) {
-                        // Check for single-quote single-line SQL
-                        if let Some(sql_start) = lines[j].strip_prefix("        sql: '") {
-                            // Single-line SQL in single quotes
-                            if let Some(sql_end) = sql_start.rfind("',") {
-                                let sql = &sql_start[..sql_end];
-                                if !database.is_empty() && !sql.is_empty() {
-                                    examples.push((id.to_string(), database.clone(), sql.to_string()));
-                                }
-                                break;
-                            }
-                        }
-                        // Check for backtick multi-line SQL
-                        else if let Some(first_line_start) = lines[j].strip_prefix("        sql: `") {
-                            // Multi-line SQL in backticks
-                            let mut sql_lines = Vec::new();
-
-                            // Check if SQL starts on the same line
-                            if !first_line_start.is_empty() && !first_line_start.ends_with("`,") && !first_line_start.ends_with("`") {
-                                // SQL starts on same line as `sql: ``, add it
-                                sql_lines.push(first_line_start);
-                            }
-
-                            // Collect remaining lines
-                            let mut k = j + 1;
-                            while k < lines.len() {
-                                let line = lines[k];
-                                if line.ends_with("`,") || line.ends_with("`") {
-                                    // Check if there's SQL on the closing line before the backtick
-                                    let line_trimmed = line.trim_end_matches("`,").trim_end_matches("`").trim_end();
-                                    if !line_trimmed.is_empty() {
-                                        sql_lines.push(line_trimmed);
-                                    }
-                                    break;
-                                }
-                                // Remove leading whitespace but preserve SQL formatting
-                                sql_lines.push(line.trim_start());
-                                k += 1;
-                            }
-
-                            if !database.is_empty() && !sql_lines.is_empty() {
-                                let sql = sql_lines.join("\n");
-                                examples.push((id.to_string(), database.clone(), sql));
-                            }
-                            break;
-                        }
-                    }
-                }
-            }
-            i += 1;
-        }
-
-        examples
-    }
-
-    #[test]
-    fn test_examples_parsing() {
-        let examples = parse_examples();
-
-        // Should have parsed examples
-        assert!(
-            examples.len() > 50,
-            "Should parse at least 50 examples, got {}",
-            examples.len()
-        );
-
-        // Verify we got the first few examples we know exist
-        let ids: Vec<&String> = examples.iter().map(|(id, _, _)| id).collect();
-        assert!(
-            ids.contains(&&"basic-1".to_string()),
-            "Should have basic-1"
-        );
-        assert!(ids.contains(&&"join-1".to_string()), "Should have join-1");
-
-        // Debug: print basic-2 SQL
-        for (id, _, sql) in &examples {
-            if id == "basic-2" {
-                eprintln!("\n=== basic-2 SQL ===\n{}\n=== end ===\n", sql);
-                break;
-            }
-        }
-    }
-
-    /// Check if an example uses known unsupported SQL features
-    fn uses_unsupported_features(sql: &str, example_id: &str) -> Option<&'static str> {
-        let sql_upper = sql.to_uppercase();
-
-        // DDL examples test CREATE TABLE with constraints (see issue #214)
-        if example_id.starts_with("ddl-") {
-            return Some("CREATE TABLE with constraints (PRIMARY KEY, UNIQUE, CHECK, etc.)");
-        }
-
-        // Check for known unsupported features
-        if sql_upper.contains("WITH ") && (sql_upper.contains(" AS (") || sql_upper.contains(" AS(")) {
-            return Some("Common Table Expressions (CTEs)");
-        }
-        if sql_upper.contains(" UNION ") || sql_upper.contains(" INTERSECT ") || sql_upper.contains(" EXCEPT ") {
-            return Some("SET operations (UNION/INTERSECT/EXCEPT)");
-        }
-        if sql_upper.contains("ALTER TABLE") || sql_upper.contains("DROP INDEX") {
-            return Some("DDL operations (ALTER TABLE/DROP INDEX)");
-        }
-        if sql.contains('|') && !sql.contains("'|'") {
-            return Some("pipe character (possibly unsupported syntax)");
-        }
-        // Aggregate functions with GROUP BY or multi-column aggregates
-        if (sql_upper.contains("COUNT(") || sql_upper.contains("AVG(") || sql_upper.contains("SUM(")
-            || sql_upper.contains("MIN(") || sql_upper.contains("MAX("))
-            && (sql_upper.contains("GROUP BY") || sql_upper.matches("COUNT(").count() > 1
-                || sql_upper.contains("AVG(") || sql_upper.contains("SUM(")) {
-            return Some("Aggregate functions with GROUP BY or multiple aggregates");
-        }
-        // Subqueries
-        if sql_upper.contains("SELECT") && sql_upper.matches("SELECT").count() > 1 {
-            return Some("Subqueries");
-        }
-        // Recursive CTEs
-        if sql_upper.contains("RECURSIVE") {
-            return Some("Recursive CTEs");
-        }
-        if sql_upper.contains("EXTRACT(") || sql_upper.contains("DATE_PART(") || sql_upper.contains("YEAR(") {
-            return Some("Date/time extraction functions");
-        }
-
-        None
-    }
-
-    #[test]
-    fn test_all_examples_execute_without_errors() {
-        let examples = parse_examples();
-
-        let mut failed_examples = Vec::new();
-        let mut skipped_examples = Vec::new();
-        let mut database_load_failures = Vec::new();
-        let mut passed_examples = 0;
-
-        for (id, db_name, sql) in &examples {
-            // Check if this example uses known unsupported features
-            if let Some(unsupported_feature) = uses_unsupported_features(sql, id) {
-                skipped_examples.push((id.clone(), unsupported_feature));
-                continue;
-            }
-
-            // Create appropriate database
-            let (mut db, db_load_error) = setup_example_database(db_name);
-
-            // If database failed to load, categorize separately
-            if let Some(load_error) = db_load_error {
-                database_load_failures.push((id.clone(), load_error));
-                continue;
-            }
-
-            // Try to execute the SQL
-            let result = execute_sql(&mut db, sql);
-
-            if let Err(error) = result {
-                failed_examples.push((id.clone(), error));
-            } else {
-                passed_examples += 1;
-            }
-        }
-
-        // Print summary
-        eprintln!("\n=== Examples Test Summary ===");
-        eprintln!("✅ Passed: {}", passed_examples);
-        eprintln!("⏭️  Skipped: {} (use unsupported SQL features)", skipped_examples.len());
-        eprintln!("⚠️  Database issues: {} (required DB failed to load)", database_load_failures.len());
-        eprintln!("❌ Failed: {} (unexpected errors)", failed_examples.len());
-        eprintln!("📊 Total: {}", examples.len());
-
-        if !skipped_examples.is_empty() {
-            eprintln!("\nSkipped examples (unsupported SQL features):");
-            let mut skipped_by_feature: std::collections::HashMap<&str, Vec<&str>> = std::collections::HashMap::new();
-            for (id, feature) in &skipped_examples {
-                skipped_by_feature.entry(*feature).or_insert_with(Vec::new).push(id.as_str());
-            }
-            for (feature, ids) in skipped_by_feature {
-                eprintln!("  {} ({}): {}", feature, ids.len(), ids.join(", "));
-            }
-        }
-
-        if !database_load_failures.is_empty() {
-            eprintln!("\nDatabase load issues (parser limitations with CREATE TABLE constraints):");
-            eprintln!("  Note: These examples require northwind/employees databases");
-            eprintln!("  Parser doesn't yet support PRIMARY KEY, NOT NULL, UNIQUE, CHECK in CREATE TABLE");
-            eprintln!("  Count: {} examples affected", database_load_failures.len());
-        }
-
-        // Report unexpected failures (but don't fail test - this is informational)
-        if !failed_examples.is_empty() {
-            eprintln!("\n⚠️  {} examples with supported features had errors:", failed_examples.len());
-            for (id, err) in &failed_examples {
-                eprintln!("  ❌ {}: {}", id, err);
-            }
-            eprintln!("\n  Note: These are examples that need table setup or have SQL compatibility issues.");
-        }
-
-        // Success message
-        if passed_examples > 0 {
-            eprintln!("\n✅ All {} examples with fully supported features passed!", passed_examples);
-        } else {
-            eprintln!("\n📝 Summary:");
-            eprintln!("   - Test infrastructure is working correctly");
-            eprintln!("   - All 73 examples are categorized (passed/skipped/database issues/errors)");
-            eprintln!("   - Main blocker for more passing tests: issue #214 (CREATE TABLE constraints)");
-            eprintln!("   - Once constraints are supported, 14+ examples will be unblocked");
-        }
-    }
-
-    #[test]
-    fn test_examples_have_database_specified() {
-        let examples = parse_examples();
-
-        let mut missing_db_examples = Vec::new();
-
-        for (id, db_name, _) in examples {
-            if db_name.is_empty() {
-                missing_db_examples.push(id);
-            }
-        }
-
-        if !missing_db_examples.is_empty() {
-            let mut error_msg = format!(
-                "\n{} examples missing database specification:\n\n",
-                missing_db_examples.len()
-            );
-            for id in &missing_db_examples {
-                error_msg.push_str(&format!("  ❌ {}\n", id));
-            }
-            panic!("{}", error_msg);
-        }
-
-        // Note: All database names are valid - northwind and employees are pre-loaded,
-        // others (empty, company, university, etc.) start as empty and examples create their own tables
-    }
 }
